requirements.txt
pydantic>=2.0.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
openai>=1.0.0
anthropic>=0.40.0
tqdm>=4.0.0
